+++ protected/humhub/config/web.php
@@ -3,7 +3,8 @@
 $config = [
     'id' => 'humhub',
     'bootstrap' => ['humhub\components\bootstrap\LanguageSelector'],
-    'defaultRoute' => '/dashboard/dashboard',
+    'homeUrl' => ['/dashboard/dashboard'],
+    'defaultRoute' => '/home',
     'layoutPath' => '@humhub/views/layouts',
     'components' => [
         'request' => [