+++ protected/humhub/docs/CHANGELOG.md
@@ -91,6 +91,7 @@ HumHub Change Log
 - Fix: Better notification compatiblity - mail views and enabled WebNotificationTarget
 - Fix #2312: Pinned post appears twice on stream 
 - Enh: Added option to show/hide deactivated user content in stream
+- Enh: Allow any url route as homepage by homeUrl array application parameter
 
 
 1.2.0-beta.2 (February 24, 2017)